fixed bugs

+++ step2_SourceTargetAdapt.py
@@ -67,8 +67,8 @@ def _do_epoch(feature_extractor, obj_cls, self_cls, source_loader,target_loader_
         img_corrects += torch.sum(cls_pred_source == class_l_source)
         
 
-    acc_cls = (img_corrects.double() / total_source_loader_num ) * 100
-    acc_rot = (self_corrects.double() / total_source_loader_num) * 100
+    acc_cls = (img_corrects / total_source_loader_num ) * 100
+    acc_rot = (self_corrects / total_source_loader_num) * 100
 
     print("Class Loss %.4f, Class Accuracy %.4f,Rot Loss %.4f, Rot Accuracy %.4f" % (class_loss.item(), acc_cls, self_loss.item(), acc_rot))
